+++ src/game/unk_091080.h
@@ -2,6 +2,6 @@
 #define _UNK_091080_H_
 #include "ultra64.h"
 
-void sub_GAME_7F091080(s32 arg0, s32 arg1, s32 arg2);
+void sub_GAME_7F091080(s8 arg0, s8 arg1, u16 arg2);
 
 #endif
